feat: add event queue infrastructure for background vote processing
- Rename checkers-batch-service to checkers-event-handler-service
- Refactor event handler into modular structure (handlers/, types, constants, utils)
- Add checkers-events-queue for async event processing
- Add vote.submitted event publishing from API route
- Add POLL_UPDATE_QUEUE binding to event handler for future use
- Create publishCheckersEvent utility with service binding for local dev
- Fix Telegram callback query error handling (safeAnswerCallbackQuery)
- Fix ObjectId validation in findCheckersVote
- Update GitHub Actions workflows for renamed service
- Update documentation (CLAUDE.md, README.md)

Note: Vote assessment still runs synchronously in API route.
The queue infrastructure is ready for future migration.
