+++ .github/workflows/deploy_secrets.yml
@@ -16,7 +16,7 @@ on:
 env:
   CHECKMATE_CHECKERS_APP_SECRET_NAMES: NEXTAUTH_SECRET,TELEGRAM_BOT_TOKEN
   CHECKERS_DB_SERVICE_SECRET_NAMES: MONGODB_CONNECTION_STRING
-  CHECKERS_BATCH_SERVICE_SECRET_NAMES: TELEGRAM_BOT_TOKEN,TELEGRAM_ADMIN_BOT_TOKEN,CHECKERS_GROUP_LINK,CHECKERS_CHAT_ID
+  CHECKERS_EVENT_HANDLER_SERVICE_SECRET_NAMES: TELEGRAM_BOT_TOKEN,TELEGRAM_ADMIN_BOT_TOKEN,CHECKERS_GROUP_LINK,CHECKERS_CHAT_ID
   CHECKERS_REMINDER_ALARM_SERVICE_SECRET_NAMES: TELEGRAM_BOT_TOKEN
   CHECKERS_WEBHOOK_SERVICE_SECRET_NAMES: TELEGRAM_BOT_TOKEN,TELEGRAM_ADMIN_BOT_TOKEN,TELEGRAM_WEBHOOK_SECRET,TYPEFORM_SECRET_TOKEN,CHECKERS_CHAT_ID,CHECKERS_GROUP_LINK,API_KEY
 
@@ -35,7 +35,7 @@ jobs:
         id: changed-workers
         run: |
           # Define all worker directories (use "." for root app)
-          ALL_WORKER_DIRS='["checkmate-checkers-app","checkers-db-service","checkers-batch-service","checkers-reminder-alarm-service","checkers-webhook-service"]'
+          ALL_WORKER_DIRS='["checkmate-checkers-app","checkers-db-service","checkers-event-handler-service","checkers-reminder-alarm-service","checkers-webhook-service"]'
 
           # Case 1: Manual trigger with force_deploy=true - deploy all workers secrets
           if [[ "${{ github.event_name }}" == "workflow_dispatch" && "${{ github.event.inputs.force_deploy }}" == "true" ]]; then
@@ -60,7 +60,7 @@ jobs:
   deploy-secrets:
     name: "Deploy worker secrets to ${{  github.ref_name == 'main' && 'production' || github.ref_name }} (${{ matrix.worker-dir }})"
     needs: process-secrets
-    if: "${{ needs.process-secrets.outputs.worker-dirs != '[]' }}"
+    if: ${{ needs.process-secrets.outputs.worker-dirs != '[]' }}
     runs-on: ubuntu-latest
     environment: "${{  github.ref_name == 'main' && 'production' || github.ref_name }}"
     strategy:
@@ -69,7 +69,7 @@ jobs:
         worker-dir:
           - checkmate-checkers-app
           - checkers-db-service
-          - checkers-batch-service
+          - checkers-event-handler-service
           - checkers-reminder-alarm-service
           - checkers-webhook-service
     timeout-minutes: 60 # Set timeout to prevent hung deployments
@@ -125,8 +125,8 @@ jobs:
             checkers-db-service)
               SECRET_NAMES="${{ env.CHECKERS_DB_SERVICE_SECRET_NAMES }}"
               ;;
-            checkers-batch-service)
-              SECRET_NAMES="${{ env.CHECKERS_BATCH_SERVICE_SECRET_NAMES }}"
+            checkers-event-handler-service)
+              SECRET_NAMES="${{ env.CHECKERS_EVENT_HANDLER_SERVICE_SECRET_NAMES }}"
               ;;
             checkers-reminder-alarm-service)
               SECRET_NAMES="${{ env.CHECKERS_REMINDER_ALARM_SERVICE_SECRET_NAMES }}"
